refactor: exclude stores and actions in account module exports
