+++ packages/desktop/components/drawers/network-config/drawers/ConfirmLedgerEvmAddressDrawer.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { DrawerTemplate } from '@components'
-    import { selectedAccount } from '@core/account'
+    import { selectedAccount } from '@core/account/stores'
     import { localize } from '@core/i18n'
     import { IIscpChainConfiguration } from '@core/network/interfaces'
     import { selectedChain } from '@core/network/stores'
