+++ packages/desktop/components/popups/AliasConfirmationPopup.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { selectedAccount, updateSelectedAccount } from '@core/account'
+    import { selectedAccount, updateSelectedAccount } from '@core/account/stores'
     import { processAndAddToActivities } from '@core/activity/utils'
     import { handleError } from '@core/error/handlers/handleError'
     import { localize } from '@core/i18n'
