+++ packages/shared/src/components/buttons/ReceiveAddressButton.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import { selectedAccount } from '@core/account'
+    import { selectedAccount } from '@core/account/stores'
     import { appSettings } from '@core/app/stores'
     import { localize } from '@core/i18n'
     import { QR, Text, FontWeight, AddressBox } from '@ui'
