+++ packages/shared/src/components/inputs/AccountInput.svelte
@@ -1,6 +1,7 @@
 <script lang="ts">
-    import { Modal, RecipientAccountSelector, AccountLabel } from '@ui'
-    import { IAccountState, selectedAccount } from '@core/account'
+    import { IAccountState } from '@core/account'
+    import { selectedAccount } from '@core/account/stores'
+    import { AccountLabel, Modal, RecipientAccountSelector } from '@ui'
 
     export let account: IAccountState = $selectedAccount
