+++ packages/shared/src/components/modals/RecipientAccountSelector.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-    import { IAccountState, selectedAccount } from '@core/account'
+    import { IAccountState } from '@core/account'
+    import { selectedAccount } from '@core/account/stores'
     import { getBaseToken } from '@core/profile/actions'
     import { visibleActiveAccounts } from '@core/profile/stores'
     import { truncateString } from '@core/utils'
