+++ packages/shared/src/components/molecules/menu-items/ToggleHiddenAccountMenuItem.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-    import { selectedAccount, setNextSelectedAccount } from '@core/account'
+    import { setNextSelectedAccount } from '@core/account/actions'
+    import { selectedAccount } from '@core/account/stores'
     import { localize } from '@core/i18n'
     import { updateActiveAccountPersistedData } from '@core/profile/actions'
     import { activeProfile, nonHiddenActiveAccounts } from '@core/profile/stores'
