+++ packages/shared/src/lib/core/account/actions/setNextSelectedAccount.ts
@@ -1,4 +1,5 @@
-import { selectedAccount, setSelectedAccount } from '@core/account'
+import { setSelectedAccount } from '@core/account/actions'
+import { selectedAccount } from '@core/account/stores'
 import { nonHiddenActiveAccounts } from '@core/profile/stores'
 import { get } from 'svelte/store'
