+++ packages/shared/src/lib/core/profile-manager/actions/events-handlers/handleTransactionInclusionEvent.ts
@@ -1,6 +1,6 @@
 import { updateParticipationOverview } from '@contexts/governance/stores'
 import { isAccountVoting } from '@contexts/governance/utils/isAccountVoting'
-import { syncVotingPower } from '@core/account'
+import { syncVotingPower } from '@core/account/actions'
 import {
     ActivityAction,
     ActivityDirection,
