+++ packages/shared/src/lib/core/wallet/actions/mintNativeToken.ts
@@ -1,5 +1,6 @@
 import { showNotification } from '@auxiliary/notification'
-import { getSelectedAccount, updateSelectedAccount } from '@core/account'
+import { getSelectedAccount, updateSelectedAccount } from '@core/account/stores'
+import { processAndAddToActivities } from '@core/activity/utils'
 import { localize } from '@core/i18n'
 import { Converter } from '@core/utils'
 import { CreateNativeTokenParams } from '@iota/wallet'
@@ -9,7 +10,6 @@ import { buildPersistedAssetFromMetadata } from '../helpers'
 import { IIrc30Metadata, IPersistedAsset } from '../interfaces'
 import { resetMintTokenDetails } from '../stores'
 import { addPersistedAsset } from '../stores/persisted-assets.store'
-import { processAndAddToActivities } from '@core/activity/utils'
 
 export async function mintNativeToken(
     maximumSupply: number,
