+++ packages/shared/src/lib/core/wallet/actions/send/signAndSendStardustTransaction.ts
@@ -1,8 +1,9 @@
-import { IAccountState, updateSelectedAccount } from '@core/account'
+import { IAccountState } from '@core/account'
+import { updateSelectedAccount } from '@core/account/stores'
+import { processAndAddToActivities } from '@core/activity/utils'
 import { updateNftInAllAccountNfts } from '@core/nfts/actions'
 import { DEFAULT_TRANSACTION_OPTIONS, OUTPUT_TYPE_NFT } from '@core/wallet/constants'
 import { Output } from '@core/wallet/types'
-import { processAndAddToActivities } from '@core/activity/utils'
 
 export async function signAndSendStardustTransaction(output: Output, account: IAccountState): Promise<void> {
     try {
